# server/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/asset": {
      "delete": {
        "operationId": "deleteAsset",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAssetDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeleteAssetResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      },
      "get": {
        "description": "Get all AssetEntity belong to the user",
        "operationId": "getAllAssets",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "updatedAfter",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "if-none-match",
            "in": "header",
            "description": "ETag of data already cached on the client",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      },
      "put": {
        "operationId": "updateAssets",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetBulkUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/download/archive": {
      "post": {
        "operationId": "downloadArchive",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetIdsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/download/info": {
      "post": {
        "operationId": "getDownloadInfo",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DownloadInfoDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/download/{id}": {
      "post": {
        "operationId": "downloadFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/ids": {
      "get": {
        "operationId": "getAssetIds",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/jobs": {
      "post": {
        "operationId": "runAssetJobs",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetJobsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/map-marker": {
      "get": {
        "operationId": "getMapMarkers",
        "parameters": [
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "fileCreatedAfter",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "fileCreatedBefore",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MapMarkerResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/memory-lane": {
      "get": {
        "operationId": "getMemoryLane",
        "parameters": [
          {
            "name": "timestamp",
            "required": true,
            "in": "query",
            "description": "Get pictures for +24 hours from this time going back x years",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemoryLaneResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/statistics": {
      "get": {
        "operationId": "getAssetStats",
        "parameters": [
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetStatsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/time-bucket": {
      "get": {
        "operationId": "getByTimeBucket",
        "parameters": [
          {
            "name": "size",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TimeBucketSize"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "personId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "timeBucket",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssetResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          },
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/time-buckets": {
      "get": {
        "operationId": "getTimeBuckets",
        "parameters": [
          {
            "name": "size",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TimeBucketSize"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "albumId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "personId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "isArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isFavorite",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TimeBucketResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          },
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/{deviceId}": {
      "get": {
        "operationId": "getUserAssetsByDeviceId",
        "parameters": [
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/{id}": {
      "put": {
        "operationId": "updateAsset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssetDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/jobs": {
      "get": {
        "operationId": "getAllJobsStatus",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllJobStatusResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Job"
        ]
      }
    },
    "/jobs/{id}": {
      "put": {
        "operationId": "sendJobCommand",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/JobName"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCommandDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Job"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.76.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AllJobStatusResponseDto": {
        "properties": {
          "backgroundTask": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "clipEncoding": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "metadataExtraction": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "objectTagging": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "recognizeFaces": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "search": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "sidecar": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "storageTemplateMigration": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "thumbnailGeneration": {
            "$ref": "#/components/schemas/JobStatusDto"
          },
          "videoConversion": {
            "$ref": "#/components/schemas/JobStatusDto"
          }
        },
        "required": [
          "thumbnailGeneration",
          "metadataExtraction",
          "videoConversion",
          "objectTagging",
          "clipEncoding",
          "storageTemplateMigration",
          "backgroundTask",
          "search",
          "recognizeFaces",
          "sidecar"
        ],
        "type": "object"
      },
      "AssetBulkUpdateDto": {
        "properties": {
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
...
      "AssetIdsDto": {
        "properties": {
          "assetIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "assetIds"
        ],
        "type": "object"
      },
...
      "AssetJobsDto": {
        "properties": {
          "assetIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "$ref": "#/components/schemas/AssetJobName"
          }
        },
        "required": [
          "assetIds",
          "name"
        ],
        "type": "object"
      },
      "AssetResponseDto": {
        "properties": {
          "checksum": {
            "description": "base64 encoded sha1 hash",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "exifInfo": {
            "$ref": "#/components/schemas/ExifResponseDto"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "livePhotoVideoId": {
            "nullable": true,
            "type": "string"
          },
          "originalFileName": {
            "type": "string"
          },
          "originalPath": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonResponseDto"
            },
            "type": "array"
          },
          "resized": {
            "type": "boolean"
          },
          "smartInfo": {
            "$ref": "#/components/schemas/SmartInfoResponseDto"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            },
            "type": "array"
          },
          "thumbhash": {
            "description": "base64 encoded thumbhash",
            "nullable": true,
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/AssetTypeEnum"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "type",
          "id",
          "deviceAssetId",
          "ownerId",
          "deviceId",
          "originalPath",
          "originalFileName",
          "resized",
          "thumbhash",
          "fileCreatedAt",
          "fileModifiedAt",
          "updatedAt",
          "isFavorite",
          "isArchived",
          "duration",
          "checksum"
        ],
        "type": "object"
      },
      "AssetStatsResponseDto": {
        "properties": {
          "images": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "videos": {
            "type": "integer"
          }
        },
        "required": [
          "images",
          "videos",
          "total"
        ],
        "type": "object"
      },
      "AssetTypeEnum": {
        "enum": [
          "IMAGE",
          "VIDEO",
          "AUDIO",
          "OTHER"
        ],
        "type": "string"
      },
...
      "DeleteAssetDto": {
        "properties": {
          "ids": {
            "example": [
              "bf973405-3f2a-48d2-a687-2ed4167164be",
              "dd41870b-5d00-46d2-924e-1d8489a0aa0f",
              "fad77c3f-deef-4e7e-9608-14c1aa4e559a"
            ],
            "items": {
              "type": "string"
            },
            "title": "Array of asset IDs to delete",
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
      "DeleteAssetResponseDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DeleteAssetStatus"
          }
        },
        "required": [
          "status",
          "id"
        ],
        "type": "object"
      },
      "DeleteAssetStatus": {
        "enum": [
          "SUCCESS",
          "FAILED"
        ],
        "type": "string"
      },
...
      "DownloadInfoDto": {
        "properties": {
          "albumId": {
            "format": "uuid",
            "type": "string"
          },
          "archiveSize": {
            "type": "integer"
          },
          "assetIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "userId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
      "DownloadResponseDto": {
        "properties": {
          "archives": {
            "items": {
              "$ref": "#/components/schemas/DownloadArchiveInfo"
            },
            "type": "array"
          },
          "totalSize": {
            "type": "integer"
          }
        },
        "required": [
          "totalSize",
          "archives"
        ],
        "type": "object"
      },
...
      "JobCommandDto": {
        "properties": {
          "command": {
            "$ref": "#/components/schemas/JobCommand"
          },
          "force": {
            "type": "boolean"
          }
        },
        "required": [
          "command",
          "force"
        ],
        "type": "object"
      },
...
      "JobName": {
        "enum": [
          "thumbnailGeneration",
          "metadataExtraction",
          "videoConversion",
          "objectTagging",
          "recognizeFaces",
          "clipEncoding",
          "backgroundTask",
          "storageTemplateMigration",
          "search",
          "sidecar"
        ],
        "type": "string"
      },
...
      "JobStatusDto": {
        "properties": {
          "jobCounts": {
            "$ref": "#/components/schemas/JobCountsDto"
          },
          "queueStatus": {
            "$ref": "#/components/schemas/QueueStatusDto"
          }
        },
        "required": [
          "jobCounts",
          "queueStatus"
        ],
        "type": "object"
      },
...
      "MapMarkerResponseDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "lat": {
            "format": "double",
            "type": "number"
          },
          "lon": {
            "format": "double",
            "type": "number"
          }
        },
        "required": [
          "id",
          "lat",
          "lon"
        ],
        "type": "object"
      },
      "MemoryLaneResponseDto": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetResponseDto"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "assets"
        ],
        "type": "object"
      },
...
      "PersonResponseDto": {
        "properties": {
          "birthDate": {
            "format": "date",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isHidden": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "thumbnailPath": {
            "type": "string"
          }
        },
        "required": [
          "birthDate",
          "id",
          "name",
          "thumbnailPath",
          "isHidden"
        ],
        "type": "object"
      },
...
      "TimeBucketResponseDto": {
        "properties": {
          "count": {
            "type": "integer"
          },
          "timeBucket": {
            "type": "string"
          }
        },
        "required": [
          "timeBucket",
          "count"
        ],
        "type": "object"
      },
      "TimeBucketSize": {
        "enum": [
          "DAY",
          "MONTH"
        ],
        "type": "string"
      },
...
      "UpdateAssetDto": {
        "properties": {
          "description": {
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    }
  }
}